Move `handle-errored-bill-run` service up one level This service was built as part of the supplementary billing engine but can be reused for all bill run types
