+++ app/services/bill-runs/supplementary/process-bill-run.service.js
@@ -10,7 +10,7 @@ const BillRunError = require('../../../errors/bill-run.error.js')
 const ChargingModuleGenerateService = require('../../charging-module/generate-bill-run.service.js')
 const FetchChargeVersionsService = require('./fetch-charge-versions.service.js')
 const { calculateAndLogTimeTaken, currentTimeInNanoseconds } = require('../../../lib/general.lib.js')
-const HandleErroredBillRunService = require('./handle-errored-bill-run.service.js')
+const HandleErroredBillRunService = require('../handle-errored-bill-run.service.js')
 const LegacyRequestLib = require('../../../lib/legacy-request.lib.js')
 const ProcessBillingPeriodService = require('./process-billing-period.service.js')
 const ReissueBillsService = require('./reissue-bills.service.js')
