+++ test/services/bill-runs/supplementary/process-bill-run.service.test.js
@@ -18,7 +18,7 @@ const DatabaseHelper = require('../../../support/helpers/database.helper.js')
 const ChargingModuleGenerateService = require('../../../../app/services/charging-module/generate-bill-run.service.js')
 const FeatureFlagsConfig = require('../../../../config/feature-flags.config.js')
 const FetchChargeVersionsService = require('../../../../app/services/bill-runs/supplementary/fetch-charge-versions.service.js')
-const HandleErroredBillRunService = require('../../../../app/services/bill-runs/supplementary/handle-errored-bill-run.service.js')
+const HandleErroredBillRunService = require('../../../../app/services/bill-runs/handle-errored-bill-run.service.js')
 const LegacyRequestLib = require('../../../../app/lib/legacy-request.lib.js')
 const ProcessBillingPeriodService = require('../../../../app/services/bill-runs/supplementary/process-billing-period.service.js')
 const ReissueBillsService = require('../../../../app/services/bill-runs/supplementary/reissue-bills.service.js')
